backend/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
python-jose==3.3.0
python-multipart==0.0.6
torch==2.4.0
torchvision==0.19.0
Pillow==9.2.0